Switch to DoubleMinor

+++ src/main/java/minispark/Scheduler.java
@@ -251,7 +251,7 @@ public class Scheduler {
         for (int i = 1; i < reduceResults.size(); ++i) {
           try {
             Method method = App.class.getMethod(function, double.class, double.class);
-            reduceResult = (int) method.invoke(null, reduceResult, reduceResults.get(i));
+            reduceResult = (double) method.invoke(null, reduceResult, reduceResults.get(i));
           } catch (Exception e) {
             e.printStackTrace();
           }
